added screenshots

+++ lib/presentation/controllers/manga/manga_controller.dart
@@ -1,4 +1,3 @@
-import 'package:collection/collection.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -16,20 +15,29 @@ class MangaController extends StateNotifier<MangaState> {
 
   MangaController({
     required MangaRepository mangaRepository,
-    required Manga? selectedManga,
     required this.favoriteMangas,
   })  : _mangaRepository = mangaRepository,
-        super(
-          MangaState(
-            isFavorite: favoriteMangas?.contains(selectedManga) ?? false,
-          ),
-        );
+        super(const MangaState());
 
-  Future<void> addToFavorite(Manga manga) =>
-      _mangaRepository.addMangaToFavorite(manga);
+  bool isMangaFavorite(Manga manga) => favoriteMangas?.contains(manga) ?? false;
 
-  Future<void> removeFromFavorite(Manga manga) =>
-      _mangaRepository.removeMangaFromFavorite(manga);
+  Future<bool> toggleFavorite(Manga manga) async {
+    state = state.copyWith(isLoading: true);
+    final isFavorite = isMangaFavorite(manga);
+    final Result<void, Object> result;
+    if (isFavorite) {
+      result = await _mangaRepository.removeMangaFromFavorite(manga);
+    } else {
+      result = await _mangaRepository.addMangaToFavorite(manga);
+    }
+    if (mounted) {
+      state = state.copyWith(
+        isLoading: false,
+        hasError: result.type == ResultType.error,
+      );
+    }
+    return !isFavorite;
+  }
 
   Future<void> updateManga(Manga manga) async {
     state = state.copyWith(isLoading: true);
@@ -54,9 +62,6 @@ final mangaControllerProvider = StateNotifierProvider.autoDispose
 
   return MangaController(
     mangaRepository: repository,
-    selectedManga:
-        repository.getFavorites().firstWhereOrNull((e) => e.id == id) ??
-            repository.getFetchedMangas().firstWhereOrNull((e) => e.id == id),
     favoriteMangas: favoriteMangas,
   );
 });
